\name{CNV.CellLine}
\alias{CNV.CellLine}
\docType{data}
\title{
R dataframe/SummarizedExperiment object with the copy number values}
\description{
R dataframe/SummarizedExperiment object with the copy number values for each probe and cell type}
\usage{data("CNV.CellLine")}
\format{
  A data frame with 100 observations on the following 100 variables.
  \describe{
    \item{\code{LOUNH91_LUNG}}{a numeric vector}
    \item{\code{T98G_CENTRAL_NERVOUS_SYSTEM}}{a numeric vector}
    \item{\code{IPC298_SKIN}}{a numeric vector}
    \item{\code{RPMI8226_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{MIAPACA2_PANCREAS}}{a numeric vector}
    \item{\code{HS695T_SKIN}}{a numeric vector}
    \item{\code{COLO704_OVARY}}{a numeric vector}
    \item{\code{SKMEL28_SKIN}}{a numeric vector}
    \item{\code{K562_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{LK2_LUNG}}{a numeric vector}
    \item{\code{NCIH2122_LUNG}}{a numeric vector}
    \item{\code{SW620_LARGE_INTESTINE}}{a numeric vector}
    \item{\code{NAMALWA_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{KMS12BM_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{G361_SKIN}}{a numeric vector}
    \item{\code{MCAS_OVARY}}{a numeric vector}
    \item{\code{X22RV1_PROSTATE}}{a numeric vector}
    \item{\code{FUOV1_OVARY}}{a numeric vector}
    \item{\code{RS411_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{OVTOKO_OVARY}}{a numeric vector}
    \item{\code{CPCN_LUNG}}{a numeric vector}
    \item{\code{OV90_OVARY}}{a numeric vector}
    \item{\code{NCIH2171_LUNG}}{a numeric vector}
    \item{\code{MOLT4_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{COLO829_SKIN}}{a numeric vector}
    \item{\code{SW579_THYROID}}{a numeric vector}
    \item{\code{RL_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{DMS79_LUNG}}{a numeric vector}
    \item{\code{PANC0403_PANCREAS}}{a numeric vector}
    \item{\code{NCIH889_LUNG}}{a numeric vector}
    \item{\code{DLD1_LARGE_INTESTINE}}{a numeric vector}
    \item{\code{HL60_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{HEYA8_OVARY}}{a numeric vector}
    \item{\code{TCCSUP_URINARY_TRACT}}{a numeric vector}
    \item{\code{HMCB_SKIN}}{a numeric vector}
    \item{\code{PANC1005_PANCREAS}}{a numeric vector}
    \item{\code{KARPAS299_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{RERFLCAI_LUNG}}{a numeric vector}
    \item{\code{NCIH524_LUNG}}{a numeric vector}
    \item{\code{SW1463_LARGE_INTESTINE}}{a numeric vector}
    \item{\code{TYKNU_OVARY}}{a numeric vector}
    \item{\code{EFM19_BREAST}}{a numeric vector}
    \item{\code{T84_LARGE_INTESTINE}}{a numeric vector}
    \item{\code{SKMEL2_SKIN}}{a numeric vector}
    \item{\code{SKMEL5_SKIN}}{a numeric vector}
    \item{\code{DU145_PROSTATE}}{a numeric vector}
    \item{\code{EFO21_OVARY}}{a numeric vector}
    \item{\code{SKMEL30_SKIN}}{a numeric vector}
    \item{\code{LN18_CENTRAL_NERVOUS_SYSTEM}}{a numeric vector}
    \item{\code{MDAMB468_BREAST}}{a numeric vector}
    \item{\code{MM1S_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{SBC5_LUNG}}{a numeric vector}
    \item{\code{RMGI_OVARY}}{a numeric vector}
    \item{\code{CHL1_SKIN}}{a numeric vector}
    \item{\code{U118MG_CENTRAL_NERVOUS_SYSTEM}}{a numeric vector}
    \item{\code{A253_SALIVARY_GLAND}}{a numeric vector}
    \item{\code{BXPC3_PANCREAS}}{a numeric vector}
    \item{\code{TCCPAN2_PANCREAS}}{a numeric vector}
    \item{\code{CAKI2_KIDNEY}}{a numeric vector}
    \item{\code{EFO27_OVARY}}{a numeric vector}
    \item{\code{NCIH2030_LUNG}}{a numeric vector}
    \item{\code{ASPC1_PANCREAS}}{a numeric vector}
    \item{\code{MFE296_ENDOMETRIUM}}{a numeric vector}
    \item{\code{MFE280_ENDOMETRIUM}}{a numeric vector}
    \item{\code{MV411_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE}}{a numeric vector}
    \item{\code{U87MG_CENTRAL_NERVOUS_SYSTEM}}{a numeric vector}
    \item{\code{MDAMB231_BREAST}}{a numeric vector}
    \item{\code{OVSAHO_OVARY}}{a numeric vector}
    \item{\code{NCIH520_LUNG}}{a numeric vector}
    \item{\code{MFE319_ENDOMETRIUM}}{a numeric vector}
    \item{\code{NUGC3_STOMACH}}{a numeric vector}
    \item{\code{NCIH446_LUNG}}{a numeric vector}
    \item{\code{RT112_URINARY_TRACT}}{a numeric vector}
    \item{\code{MELHO_SKIN}}{a numeric vector}
    \item{\code{UACC257_SKIN}}{a numeric vector}
    \item{\code{EFM192A_BREAST}}{a numeric vector}
    \item{\code{SW900_LUNG}}{a numeric vector}
    \item{\code{MELJUSO_SKIN}}{a numeric vector}
    \item{\code{A2058_SKIN}}{a numeric vector}
    \item{\code{A2780_OVARY}}{a numeric vector}
    \item{\code{OVCAR4_OVARY}}{a numeric vector}
    \item{\code{L33_PANCREAS}}{a numeric vector}
    \item{\code{HS944T_SKIN}}{a numeric vector}
    \item{\code{VMRCRCZ_KIDNEY}}{a numeric vector}
    \item{\code{DBTRG05MG_CENTRAL_NERVOUS_SYSTEM}}{a numeric vector}
    \item{\code{HCT116_LARGE_INTESTINE}}{a numeric vector}
    \item{\code{HS766T_PANCREAS}}{a numeric vector}
    \item{\code{SKMEL24_SKIN}}{a numeric vector}
    \item{\code{CAOV3_OVARY}}{a numeric vector}
    \item{\code{PANC0203_PANCREAS}}{a numeric vector}
    \item{\code{DOV13_OVARY}}{a numeric vector}
    \item{\code{HT29_LARGE_INTESTINE}}{a numeric vector}
    \item{\code{AZ521_STOMACH}}{a numeric vector}
    \item{\code{C32_SKIN}}{a numeric vector}
    \item{\code{CAL851_BREAST}}{a numeric vector}
    \item{\code{NCIH1975_LUNG}}{a numeric vector}
    \item{\code{A549_LUNG}}{a numeric vector}
    \item{\code{SCABER_URINARY_TRACT}}{a numeric vector}
    \item{\code{BT549_BREAST}}{a numeric vector}
    \item{\code{HCT15_LARGE_INTESTINE}}{a numeric vector}
  }
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}
\examples{
data(CNV.CellLine)
## maybe str(CNV.CellLine) ; plot(CNV.CellLine) ...
}
\keyword{datasets}
