\name{CNV.Sample}
\alias{CNV.Sample}
\docType{data}
\title{
R dataframe object with gene-centric copy number variation for the samples}
\description{
This R dataframe object has gene-centric copy number variation for the samples. The rows are the genes and columns are the samples.}
\usage{data("CNV.Sample")}
\format{
  A data frame with 5 observations on the following 79 variables.
  \describe{
    \item{\samp{TCGA-OR-A5JT-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K6-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K3-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JI-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K9-01A}}{a numeric vector}
    \item{\samp{TCGA-OU-A5PI-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JV-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JE-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5L8-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5J1-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JC-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LE-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LL-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JZ-01A}}{a numeric vector}
    \item{\samp{TCGA-PK-A5HA-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K1-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5L6-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LG-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5J2-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5J5-01A}}{a numeric vector}
    \item{\samp{TCGA-PA-A5YG-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5J6-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5L9-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LT-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LA-01A}}{a numeric vector}
    \item{\samp{TCGA-P6-A5OG-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JG-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LD-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LK-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LO-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LM-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5KT-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5KV-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K4-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JD-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LJ-01A}}{a numeric vector}
    \item{\samp{TCGA-PK-A5H8-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JJ-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5KY-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K2-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5KU-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JP-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JW-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5L3-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5KX-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JR-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K8-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JF-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LR-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5L5-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JX-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LS-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JY-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5J8-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LC-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JB-01A}}{a numeric vector}
    \item{\samp{TCGA-P6-A5OF-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JL-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JA-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LN-01A}}{a numeric vector}
    \item{\samp{TCGA-PK-A5HB-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5KZ-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JK-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5KO-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5L4-01A}}{a numeric vector}
    \item{\samp{TCGA-PK-A5H9-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5KW-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5J7-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5J9-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JQ-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JS-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K5-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JM-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LH-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5JO-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LB-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5K0-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5J3-01A}}{a numeric vector}
    \item{\samp{TCGA-OR-A5LP-01A}}{a numeric vector}
  }
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}
\examples{
data(CNV.Sample)
## maybe str(CNV.Sample) ; plot(CNV.Sample) ...
}
\keyword{datasets}
