\name{Deconv_expression}
\alias{Deconv_expression}
\docType{data}
\title{
Deconvoluted Expression Profile
}
\description{
This is a deconvoluted expression profile that can be generated by the RunDeconvExpr() function.
}
\usage{data("Deconv_expression")}
\format{
  The format is:
 num [1:5, 1:3] 10.217 0.107 35.665 1.392 0.232 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:5] "TSPAN6" "TNMD" "DPM1" "SCYL3" ...
  ..$ : NULL
}
\examples{
data(Deconv_expression)
## maybe str(Deconv_expression) ; plot(Deconv_expression) ...
}
\keyword{datasets}
