\name{Deconv_meth_gene}
\alias{Deconv_meth_gene}
\docType{data}
\title{
Gene Centric Methylation Profile
}
\description{
This R dataframe object is a gene centric methlation profile that can be generated by the probeToGeneFun() function.
}
\usage{data("Deconv_meth_gene")}
\format{
  A data frame with 75 observations on the following 3 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
  }
}
\examples{
data(Deconv_meth_gene)
## maybe str(Deconv_meth_gene) ; plot(Deconv_meth_gene) ...
}
\keyword{datasets}
