\name{Deconv_proportions}
\alias{Deconv_proportions}
\docType{data}
\title{
Deconoluted proportions
}
\description{
This R datafrane object has the deconvoluted proportions of cell types. This can be generated by the RunDeconvMethylFun() function.
}
\usage{data("Deconv_proportions")}
\format{
  The format is:
 num [1:80, 1:3] 0.31 0.55 0.86 0.58 0.73 0.14 0.49 0 0 0.67 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:80] "TCGA-OU-A5PI-01A" "TCGA-OR-A5K8-01A" "TCGA-OR-A5JX-01A" "TCGA-OR-A5LT-01A" ...
  ..$ : NULL
}
\examples{
data(Deconv_proportions)
## maybe str(Deconv_proportions) ; plot(Deconv_proportions) ...
}
\keyword{datasets}
