\name{pat_de_genes}
\alias{pat_de_genes}
\docType{data}
\title{
A R dataframe object with a list of the samples DE genes.
}
\description{
A R dataframe object with a list of the samples DE genes. This file is produced in the pipeline.
}
\usage{data("pat_de_genes")}
\format{
  A data frame with 5716 observations on the following variable.
  \describe{
    \item{\code{gene}}{a character vector}
  }
}
\examples{
data(pat_de_genes)
## maybe str(pat_de_genes) ; plot(pat_de_genes) ...
}
\keyword{datasets}
