\name{Reference.Methylation.Markers}
\alias{Reference.Methylation.Markers}
\docType{data}
\title{
A vector containing DNA methylation markers(important probes)}
\description{
This file is a vector containing DNA methylation markers(important probes)of different cell types that are used in deconvolution purpose.}
\usage{data("Reference.Methylation.Markers")}
\format{
  A vector containing DNA methylation markers.
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}
\examples{
data(Reference.Methylation.Markers)
## maybe str(Reference.Methylation.Markers) ; plot(Reference.Methylation.Markers) ...
}
\keyword{datasets}
